requirements.txt
smolagent
streamlit